requirements-test.txt
# Test dependencies for DockSafe
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-flask>=1.3.0
pytest-xdist>=3.3.0
coverage>=7.0.0
factory-boy>=3.3.0
faker>=19.0.0
freezegun>=1.2.0
requests-mock>=1.11.0
